Added relevant validation for all controller logic

+++ src/controllers/appartments.controller.js
@@ -28,9 +28,10 @@ module.exports = {
     createAppartment: (req, res, next) => {
         console.log("createAppartment called");
         const appartment = req.body;
+        const userId = req.userId;
 
         // Check if postalcode is valid
-        if(postalcodeValidator.test(user.PostalCode) != true)
+        if(postalcodeValidator.test(appartment.PostalCode) != true)
         {
           errorObject = {
             message : 'Postal Code is not valid!',
@@ -41,7 +42,9 @@ module.exports = {
           return;
         }
 
-        const query = `INSERT INTO Apartment VALUES('${appartment.Description}', '${appartment.StreetAddress}', '${appartment.PostalCode}', '${appartment.City}', ${appartment.UserId} )`;
+        const query = `INSERT INTO Apartment VALUES('${appartment.Description}', '${appartment.StreetAddress}', '${appartment.PostalCode}', '${appartment.City}', ${userId} )`;
+
+        console.log(query);
 
         database.dbQuery(query, (err, rows) => {
 
@@ -61,7 +64,7 @@ module.exports = {
         console.log("getAppartment called");
 
         const id = req.params.id;
-        const query = `SELECT * FROM Apartment WHERE Apartment.ApartmentId=${id}`;
+        const query = `SELECT * FROM Apartment INNER JOIN DBUser ON Apartment.UserId = DBUser.UserId INNER JOIN Reservation ON Reservation.ApartmentId = Apartment.ApartmentId WHERE Apartment.ApartmentId=${id}`;
 
         database.dbQuery(query, (err, rows) => {
             if(err){
@@ -72,7 +75,19 @@ module.exports = {
                 next(errorObject);
             }            
             if(rows){
-                res.status(200).json({result: rows});
+
+                // Check if appartment exists
+                if(rows.length == 0)
+                {
+                    const errorObject = {
+                        message: 'Appartment does not exist!',
+                        code: 404
+                    }
+                    next(errorObject);
+                }
+                else{
+                    res.status(200).json({result: rows});
+                }
             }
         });
 
@@ -86,6 +101,18 @@ module.exports = {
 
         const appartment = req.body;
         const id = req.params.id;
+
+        // Check if postalcode is valid
+        if(postalcodeValidator.test(appartment.PostalCode) != true)
+        {
+          errorObject = {
+            message : 'Postal Code is not valid!',
+            code : 500
+          } 
+
+          next(errorObject);
+          return;
+        }
 
         const query = `UPDATE Apartment SET Apartment.Description = '${appartment.Description}', Apartment.StreetAddress = '${appartment.StreetAddress}', Apartment.PostalCode = '${appartment.PostalCode}', Apartment.City = '${appartment.City}' WHERE Apartment.ApartmentId = ${id};`;
 
@@ -109,9 +136,13 @@ module.exports = {
         console.log("deleteAppartment");
 
         const id = req.params.id;
-        const query = `DELETE FROM Apartment WHERE Apartment.ApartmentId=${id}`;
+        const callerUserId = req.userId;
+        const queryDelete = `DELETE FROM Apartment WHERE Apartment.ApartmentId=${id}`;
+        const queryCheck = `SELECT * FROM Apartment WHERE Apartment.ApartmentId=${id} AND Apartment.UserId = ${callerUserId}`
 
-        database.dbQuery(query, (err, rows) => {
+        console.log(queryCheck);
+
+        database.dbQuery(queryCheck, (err, rows) => {
 
             if(err){
                 const errorObject = {
@@ -120,17 +151,51 @@ module.exports = {
                 }
                 next(errorObject);
             }else{
-                res.status(200).json({});
+                if(rows.length == 0)
+                {
+                    const errorObject = {
+                        message: 'You are not authorized to delete apartment or object does not exist.',
+                        code: 401
+                    }
+                    next(errorObject);
+                }
+                else{
+                    database.dbQuery(queryDelete, (err, rows) => {
+
+                        if(err){
+                            const errorObject = {
+                                message: 'Error',
+                                code: 500
+                            }
+                            next(errorObject);
+                        }else{
+                            res.status(200).json({});
+                        }
+                    })
+                }
             }
         })
+
+
     },
 
     createReservation: (req, res, next) => {
 
         const id = req.params.id;
         const reservation = req.body;
+        const userId = req.userId;
 
-        const query = `INSERT INTO Reservation VALUES(${id}, '${reservation.StartDate}', '${reservation.EndDate}', '${reservation.Status}', ${reservation.UserId});`;
+        const query = `INSERT INTO Reservation VALUES(${id}, '${reservation.StartDate}', '${reservation.EndDate}', '${reservation.Status}', ${userId});`;
+
+        // Check dates
+        if(Date.parse(reservation.EndDate) < Date.parse(reservation.StartDate))
+        {
+            const errorObject = {
+                message: 'Start date is not before end date.',
+                code: 500
+            }
+            next(errorObject);
+        }
 
         console.log(query);
 
@@ -138,7 +203,7 @@ module.exports = {
 
             if(err){
                 const errorObject = {
-                    message: 'Error',
+                    message: 'Database error.',
                     code: 500
                 }
                 next(errorObject);
@@ -187,7 +252,18 @@ module.exports = {
                 }
                 next(errorObject);
             }else{
-                res.status(200).json({Result: rows});
+                if(rows == 0)
+                {
+                    const errorObject = {
+                        message: 'Error: Reservation not found.',
+                        code: 404
+                    }
+                    next(errorObject);       
+                }
+                else
+                {
+                    res.status(200).json({Result: rows});
+                }
             }
         })        
     },
@@ -199,10 +275,23 @@ module.exports = {
         const id = req.params.id;
         const rid = req.params.rid;
         const Registration = req.body;
+        const userId = req.userId;
 
-        const query = `UPDATE Reservation SET Reservation.Status = '${Registration.Status}' WHERE Reservation.reservationId=${rid} AND Reservation.ApartmentId=${id};`;
+        const queryUpdate = `UPDATE Reservation SET Reservation.Status = '${Registration.Status}' WHERE Reservation.reservationId=${rid} AND Reservation.ApartmentId=${id};`;
+        const queryCheck = `SELECT * FROM Reservation WHERE Reservation.UserId =${userId} AND Reservation.ReservationId = ${rid}`;
 
-        database.dbQuery(query, (err, rows) => {
+        // Check if status is valid
+        if( !(Registration.Status == "INITIAL") || (Registration.Status == "ACCEPTED") || (Registration.Status == "NOT_ACCEPTED") )
+        {
+            const errorObject = {
+                message: 'Status is not correct.',
+                code: 500
+            }
+            next(errorObject);
+        }
+
+        // Check if reservation is made by user Id
+        database.dbQuery(queryCheck, (err, rows) => {
 
             if(err){
                 const errorObject = {
@@ -211,7 +300,30 @@ module.exports = {
                 }
                 next(errorObject);
             }else{
-                res.status(200).json({});
+
+                if(rows == 0)
+                {
+                    const errorObject = {
+                        message: 'You are not authorized to update the reservation or object does not exist.',
+                        code: 401
+                    }
+                    next(errorObject);
+                }
+                else{
+
+                    database.dbQuery(queryUpdate, (err, rows) => {
+
+                        if(err){
+                            const errorObject = {
+                                message: 'Error',
+                                code: 500
+                            }
+                            next(errorObject);
+                        }else{
+                            res.status(200).json({});
+                        }
+                    })
+                }
             }
         })
     },
@@ -222,8 +334,9 @@ module.exports = {
 
         const id = req.params.id;
         const rid = req.params.rid;
-        
-        const query = `DELETE FROM Reservation WHERE Reservation.reservationId = ${rid} AND Reservation.apartmentId = ${id}`;
+        const userId = req.userId;
+
+        const query = `DELETE FROM Reservation WHERE Reservation.reservationId = ${rid} AND Reservation.apartmentId = ${id} AND Reservation.UserId = ${userId}`;
 
         database.dbQuery(query, (err, rows) => {
